inkforge-lite/requirements.txt
flask>=2.0
flask-cors>=4.0
pywebview>=4.0
requests>=2.28
fastmcp>=2.0
